requirements.txt
torch
torchvision
transformers
gradio
numpy
pandas
pillow
scipy
supervision
timm
einops
bitsandbytes
accelerate